ASSIGNED - bug 1586: Drops for regions and aggregation are badly designed https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1586
Updated the region drops to be linked to show region parents.  Moved the aggregation promise to be pointed to the field that is annotated, not by the regions that are mapped into.

All the model drop links are made directly in the RegionRules checkers now, and not in the computeBasedOnAST() methods in the InRegionPromiseDrop and AggregatePromiseDrop classes.  It was very confusing having the links made there.

Updated RegionModel drop to show the "static" qualifier.

This changes are exploratory, and reveal defects in the existing drop-sea framework.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@12745 d02ae147-ee12-4a35-9452-91c89e8b2592