RESOLVED - bug 1588: Lock analysis should use Effect analysis results, and not duplicate the work of aggregated state https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1588
Added public elaborateEffect() method to EffectsVisitor.

Extracted a big chunk of LockUtils.getLocksForMethodAsRegionRef() into LockUtils.getLockFromEffect().  Had to change one thing: if an instance lock is needed, and the target from which it is generated comes for BCA elaboration, we use the VariableUseExpression that was elaborated and not the expression that is present in the target.  This solves 2 issues:
(1) Lock processing doesn't use VariableDeclarators 
(2) Most of the time we really do want the variable that was used, so that we can match it up with the variable used to acqurie the lock.

Added getLocksForDirectRegionAccess() to LockUtils.  THis method takes an effect, elaborates it if possible, and then uses getLockFromEffect() to get the needed locks.

Changed LockVisitor.visitArrayRefExpression(), LockVisitor.visitFieldRef(), and LockVisitor.visitVariableDeclarator() to used getLocksForDirectRegionAccess().


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@12773 d02ae147-ee12-4a35-9452-91c89e8b2592