Removed use of LockUtils.getLockForRegionRef() from LockVisitor.deferencesSafeObject().
This changes causes warnings to be issued more broadly than they were before, because aggregation isn't being properly accounted for.  The correct thing to do is to move to the effects and CFG based lock analysis, and then issue warnings based on the effects being used to generate locks.  See bug 1512.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@12809 d02ae147-ee12-4a35-9452-91c89e8b2592