ASSIGNED - bug 1623: Retire @SingleThreaded annotation https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1623
It works fine for intrinsic locks.  There are some issues with field/instance initializers and JUC locks: some fields accesses are listed as unprotected that actually are projected because the flow analysis is querying the wrong flow unit for analysis results.

Analysis now does nothing with @SingleThreaded annotations.  Instead when looking at a constructor it looks for various combinations of other annotations and tries to prove the constructor is single threaded.  Constructors that are so proven, are listed under the appropriate lock promise drops in the Concurrency heading.

An access in a single-threaded constructor now shows in the supporting information whether the constructor that contains it is single-threaded or not.



git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@13791 d02ae147-ee12-4a35-9452-91c89e8b2592