FIXED - bug 1631: Update Analyses to Handle Instance Initialization properly https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1631 - Added InstanceInitializationVisitor - Deprecated InstanceInitVisitor - Only LockVisitor is updated to use InstanceInitializationVistor - Corrected LockExpressionVisitor too
Fixed bug in edu.uwm.cs.fluid.java.control.JavaTransfer.runClassInitializer: need to set the label list properly.

Added crude mechanism to JavaTransfer and Backwared/ForwardAnalysis (in UWM code) to access the sub analysis that is created by runClassInitializer.  This technique still requires a lot of type casts.  Should try to improve this in the future.

Updated IntrinsicLockAnalysis, MustHoldAnalysis, and MustReleaseANalysis to have queries and to handle the constructor context correctly.  Updated LockVisitor to use MustHOld and MustRelease analysis properly and to get the initializer sub analysis when appropriate.

This fixes the original problem that I started to try to fix before Christmas; Anlaysis results of JUC locks in initializers would report that the access was unprotected but show supporting information that showed the lock was held.  (This happened in RetireSingleThreadedJUC)


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@13949 d02ae147-ee12-4a35-9452-91c89e8b2592