Updated IntrinsicLockAnlaysis.Query to get the correct subanalysis for instance initialization.
Updated a the UWM flow analysis framework to avoid a bunch of type casts and other silliness involving the types of Lattice, FlowAnalysis, and JavaTransfer objects by adding additional type parameters.  I'm not sure the added parameters contribute much to clarity, but they should prevent stupid errors in subclasses.  I couldn't put tight enough constraints on the parameters to JavaBackwardTransfer and JavaForwardTransfer because it would have required the ability to refer to the SELF type, which Java sadly doesn't provide.  

Also eliminating some instanceof checks in FlowAnalysis and JavaTransfer by adding additional abstract methods and putting final implementations of the methods in BackwardAnalysis, ForwardAnalysis, and JavaBackwardTransfer, and JavaForwardTransfer.

Further refined the general pattern for creating a flow analysis:
- Extend IntraproceduralAnalysis
- Define nested interfaces extending from AnalysisQuery for each type of query
- Define a nested transfer function
- Define a nested FlowAnlaysis extension
- Transfer function handles caching and creation of the instance initialization subanalysis.

- 

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@13971 d02ae147-ee12-4a35-9452-91c89e8b2592