ASSIGNED - bug 1642: Rewrite BCA as a UWM analysis https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1642
Created a new class LocalVariableDeclarations that encapsulates the visitor that was in BindingContext for fining local variables.  Added an additional visitor and driver to search upwards/outwards for additional variable declarations that may be visible.  It's smart enough to not include any variable declarations that occur after the nested class.

Still need to update BindingContext to take advantage of the added information.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/jsure-client-eclipse@14102 d02ae147-ee12-4a35-9452-91c89e8b2592