RESOLVED - bug 1667: Qualified lock receivers complained about in Modeling Problems https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1667
Added qualified receivers to @GuardedBy.  This doesn't really do anything because we don't support them in analysis.  

I had to change the LockRules scrubbing so that the warning message about qualified receivers is generated as a warning drop instead of going to the Modeling Problems.

Outstanding cases for @GuardedBy:

** itself : For reference fields only; the object to which the field refers.

Never been sure what this means exactly.  I assume it has something to do with the aggregation or using the referenced object as a lock to protect itself, but I don't' have an example of its use.

** method-name() : The lock object is returned by calling the named nil-ary method.

This is impossible to support unless we change our annotations.  We can introduce a @ReturnsLock annotation on the named method, and introduce a @RegionLock annotation on the class that both use the same lock name.  The problem is that we don't know what the lock expression is to use in the @RegionLock annotation.  This use of @GuardedBy completely abstracts away the identity of the lock expression.  Whereas our requirement to the name the lock makes it necessary to explicitly write down the lock expression.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@14558 d02ae147-ee12-4a35-9452-91c89e8b2592