ASSIGNED - bug 1645: How to represent variables declared out of scope https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1645
Trying a BCA-specific fix.  BCA already has in the list of "ignored" variables the declarations of the variables external to the class.  Added a list of flags that works parallel to this list that marks which ignored variables are external variables (and which are primitives).  Added a special IRNode used to mark a reference to an external variable.  localObjects() now returns a singleton set of this special node if the reference is to an external variable.  Clients of analysis can/must test for this node and respond accordingly.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@14563 d02ae147-ee12-4a35-9452-91c89e8b2592