ASSIGNED - bug 1638: Try to eliminate code repetition in the the flow analysis implementations https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1638
Created the interface SubAnalysisFactory.  Created the abstract class AbstractCachingSubAnalysisFactory that handles caching of subanalyses.  Updated JavaTransfer to be parameterized (by type and by instance) with a SubAnalysisFactory.  The createAnalysis() method has been removed from JavaTransfer and moved to SubAnalysisFactory.

This still doesn't solve the problem of needing to declare getSubAnlaysis() methods in all the FlowAnalysis and JavaTransfer subclasses uses by an actual analysis implementation.  Will have to think about how to do that.

But this does consolidate the caching and management of subanalysis objects, and allows for different technqiues (such as those that will be necessary for side-effecting flow analyses) to be used easily.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@14647 d02ae147-ee12-4a35-9452-91c89e8b2592