ASSIGNED - bug 1681: Change flow analysis queries to be lazy https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1681
Changec AbstractJavaFlowAnalysisQuery to be lazy about the creation of the underlying Analysis objects.  This prevents flow analysis from being kicked off just because a query is created.  Now we wait until someone (if ever) asks the query for a result.

Had to add a getAnalysisThunk() method to Intraprocedural analysis, and add some Thunk classes to com.surelogic.util.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@14841 d02ae147-ee12-4a35-9452-91c89e8b2592