RESOLVED - bug 1681: Change flow analysis queries to be lazy https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1681
Got rid of ugly query management in LockVisitor that was designed to prevent flow analysis from running the case when it is obvious that JUC locks are not used.  THis is handled by the queries being lazy now.  The one extra case that needs to be managed in LockVisitor is calls to get the held JUC locks.  In this case we now use a dummy query MustHoldAnalysis.EMPTY_HELD_LOCKS_QUERY that always returns that no locks are held.  

Got rid of the ugly ctxtControllingDecl field from LockVisitor by doing this.

Refined JavaFlowAnalysisQuery.getSubAnalysisQuery to use JavaFlowAnalysisQuery as the return type instead of AnalysisQuery.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@14845 d02ae147-ee12-4a35-9452-91c89e8b2592