ASSIGNED - bug 1678: Minor Analysis Tasks https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1678 RESOLVED - bug 1680: EffectsVisitor doesn't handle constructor calls or anonymous class expressions correctly https://fluid.surelogic.com/cgi-bin/bugzilla3/show_bug.cgi?id=1680
Fixed 1680 and 1682.  Changed Effects.getEffectsQuery() to take a BCA query object.  Used to be that Effects had a BCA object and tried to generated the query, but this doesn't work because the appropriate context for sub analyses is lost.  Changes cascade up into the other analyses.  Updated LockExpressions.LockExpressionsVisitor to maintain a BCA query.  

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/jsure-client-eclipse@14875 d02ae147-ee12-4a35-9452-91c89e8b2592