In TRolesFirstPass.java, now use JavaSemanticsVisitor to build the TRStaticXXX tree. This ensures that we visit the correct parts of the CU in the correct order. Especially matters for class-init code.
We no longer make any distinction between TRoleNames and TRoleNameModels. So I nuked the TRoleName class and adjusted all callers to use TRoleNameModel directly.

Added code to the TRoleRename pass that checks each Class for any ThreadRoleIncompatible annotations that were processed and then linked up as direct dependents of LOCAL  TRoleNameModels. The code that builds the Incompatibility expressions expects to find all such annotations as direct dependents of the CANONICAL models (that is, they should be direct dependents of the actual defining TRoleNameModel). We meet this expectation by adding the TRoleIncompatibleDrops to the dependents of the canonicalNameModel.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@14957 d02ae147-ee12-4a35-9452-91c89e8b2592