Fixed problems caused by type parameterization:
Changed the GlobalLockModel to use type erasure when getting ClassRecords.

Changed the effects inference to in EffectsAnalysis to degrade uses of type formals to java.lang.Object.

Fixed misuse of areDirectlyRelated() in AbstractTarget. Was only meant to be used by the overlap methods, but was also being used by the checkTarget methods.  These now use a new isAncestorOf() predicate.

Replaces uses of isSubTypeOf with isRawSubTypeOf.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@15950 d02ae147-ee12-4a35-9452-91c89e8b2592