Changed uniqueness analysis module to not cache the promise records.  THere was no reason to do so.
Fixed JavaComponentFactory to actually be thread safe.  getComponent was not threadsafe, and could allow  a component to be created twice, which messed up the control flow graph.  THis was causing problems with TestUniquerReturnOnConstructor, and was the reason it sometimes failed, sometimes passed.

Removed commented-out code in other places.

Chnaged ComponentFactory to be an interface.


git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@17260 d02ae147-ee12-4a35-9452-91c89e8b2592