Updated side-effecting uniqueness analysis to time out.
Changed so that we no longer have StoreLattice recording which unique fields are are read/written by a method, and we get that from a query.  This was stupid, actually, because we already knew this information from the tree walk we use to determine which methods should be analyzed.  Changed so that we use the tree walk information directly because in the case of analysis time out, we aren't be able to get the query results from the analysis and have to get that information in the uniqueness analysis module any how.

- Enabled time out for side-effecting
   - when time out, we add the time out result to the following drops
     + Control flow drop of timed-out method
     + Borrowed drops of each borrowed parameter of the method
     + Unique drop of the return node, if present




git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@18098 d02ae147-ee12-4a35-9452-91c89e8b2592