ASSIGNED - bug 1716: Update effects analysis https://fluid.surelogic.com/bugzilla3/show_bug.cgi?id=1716
Updated TargetElaborator to handle @ReadOnly references: they are replaced with effects on Object:All.

Requires TargetElaboror to call getAllRegion(), a non-static method, so the class ahd to be changed to not be static.  This created a cascade of changes that required LockVisitor to now have an Effects field that it ultimately uses to pass down to a TargetElaborator.




git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@18795 d02ae147-ee12-4a35-9452-91c89e8b2592