ASSIGNED - bug 1716: Update effects analysis https://fluid.surelogic.com/bugzilla3/show_bug.cgi?id=1716
Updated effects to deal with @Borrowed @ReadOnly fields.  This error is detected in the TargetElaborator, but we don't always want the elaborator to report errors, only when we are assuring effects; not when we are getting effects for the use of other analyses.  Changed the target elaborator to take an ElaborationCallback that is invoked when a write efffect on an object reference by @Borrowed @ReadOnly field is detected.  Most cases just use a NullCallback.  The EffectsAnalysis provides a callback that reports results.

Adding the callback caused changed to trickle up the call tree.

git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/fluid@18814 d02ae147-ee12-4a35-9452-91c89e8b2592