Fixed uniqueness analysis to use the correct receiver when inside an instance or field initializer of a class.
Renamed transferUseQualifiedRcvr() to transferUseQualifiedReceiver()

Added transferUseReceiver() to JavaEvaluationTransfer.

Updated analyses to use transferUseReceiver().

Fixed side-effecting uniqueness to get the receiver from the flow unit
being analyzed unless the this expression is inside a field init or
class init inside an anonymous class expression or enumeration class
expression.  In those cases it gets the receiver from the init method of
the inner class.


Former-commit-id: 9649723617f99ea85cb8e22aa160390dfc59afcb