ASSIGNED - bug 1728: Parameterized types confuse Immutable assurance https://fluid.surelogic.com/bugzilla3/show_bug.cgi?id=1728
Created a new inner class in TypeUtils: UpperBoundGetter.  This replaces
the private method from TypeBasedMayAlias.getUpperBound().  It wraps a
type environment and reference to the java.lang.Object IjavaType object.
Changed TypeBasedMayAlias and ImmutableProcessor to use it.  

This fixed the problem that tim describes in the bug report in
ImmutableProcessor.

But ThreadSafeProcessor and ContainableProcessor still have similar
problems.


Former-commit-id: dc1f6d6dd4c358c2405126b19b686e8303dfe799