ASSIGNED - bug 1721: Apply kludge for read-only iterators https://fluid.surelogic.com/bugzilla3/show_bug.cgi?id=1721
Basic version of this is working.  I even add a new piece of
TargetEvidence indicating that the write was turned into a read.

Problems: using simple name matching to find the iterator() and remove()
methods.  Need to acutally look for the correct Iterable and Iterator
methods using he binder.

Also need to expand to work with ListIterator and it's add and set
methods.

Former-commit-id: 464fedb6e1a4c35eb87a089b7c938a93b3225cf2