ASSIGNED - bug 1731: Test all the error messages in unique + from https://fluid.surelogic.com/bugzilla3/show_bug.cgi?id=1731
Hacked up the uniqueness analysis module so that results for @ReadOnly
and @Immutable things are shown.  Didn't do this in the detail that was
done for other items because I just want to see any errors messages for
this for testing purposes.  The whole UniquenessAnlysisModule is going
to be replaced when this is converted to side-effecting result
reporting.


Former-commit-id: 74dae28652b90c9696a163d11629ba6c08514ac8