Fixed UAM so that methods that call methods and pass them objects whose type is @Immutable are marked as interesting.  Fixed the chain of evidence so that Immutable types whose instances are passed to methods depend on the control-flow of the methods containing those calls.
This was needed because uniqueness analysis does interesting work when
an actual parameter is a "value type", right now triggered when the type
is @Immutable.  Which out the changes above, we can skip over methods
that contain these interesting calls, and we didn't have any place to
hand the results.


Former-commit-id: cdab714458af819b1cebdc71b996f17f25c8012e