Improved unique+from to not check the whether a called method modifies the contents of parameter when the parameter is a VALUE object (that is, an instance of a @Immutable type).  This check was silly because an @Immutable object cannot be modified.  This check is only for immutable refs, where the object being reference might be otherwise mutable.
Having the check on @Immutable value objects besides being unneccessary
caused the user to have to add lots of effects annotations.

Fixed the above by making opCheckMutable use the same hack as
opCheckTopState. Now it turns out the original hack in opCheckTopState
was not correct.  Both now do the following:

When the state of the reference is IMMUTABLE, we check all the objects
that could be referenced, and if each one is either a VALUE object or a
sharable object (state <= SHARED), the reference can be treated as
SHARED.

Former-commit-id: d02c7186cb1d274b0b12c2fbfe6c519b2f830eac