Improved the error reporting from opGet() to use a special message to for errors that arise from transferNesterClassUse().  The error indicates that the value is read while caching the values of external variables used by the local class.
Fixed transferCall() to use the outer object specifier that encloses the
method call as the "source operation" for side-effects, if it exists,
because this gives more useful results.  

Had to change considerDeclaredEffects() as a result: Used to try to find
the outer object specifier directly, but now it will have as the srcOp.


Former-commit-id: 6d68f2d5c6e3c4e4c7aeae5362862f89c33ae9f0