Fixed public bug 97.  The control flow graph and transfer functions didn't handle the initialization of enum consants correct.  Actually the transfer functions trhew exceptions on enum constant declarations.  But I learned the CFG for enum constant declarations didn't evaluate in the correct order becuase I had no subcomponent of the EnumConstantDeclaration.  But I coudln't add one because I had no child syntax node to create one around.
I had Edwin create ImpliedEnumConstantIntiailization nodes for
EnumConstantDeclaration nodes that I can use for creating subcomponents.
 Had to do two things to fix the graph:

(1) The component created for the ImpledEnumConstantIntialization uses a
special ImpliedSubcomponet that returns children of the
EnumConstantDeclaration.  This is sleazy and confused things, so I also
had to

(2) Make EnumConstantDeclaration (op) use a special Component that
forwards requests for subcomponents to the Component used by the
ImpliedEnumConstnatInitation node.  

Fixed the transfer fucntions by adding the following methods:
(1) JavaTransfer.transferEnumConstantDeclaration
(2) JavaTransfer.transferImpliedNewExpression
(3) JavaEvaluationTransfer.transferImpliedEnumConstantInitialization

(1) Should be handled as a special kind of field initialization
(2) Should be handled as a special kind of constructor call
(3) Just passes the value through, and shouldn't normally need to be
overridden.

Had to fix a bunch of transfer functions and other methods to hande
EnumConstantClassDeclarations in the same manner as
AnonClassExpressions.

Fixed JavaEvaluationTransfer to transfer for
EnumConstantClassDeclaration as both a field init and a nested class
declaration.

Things seem to be working.  But still need to compare execution (via
invoked transfer function methods) of EnumConstantDeclarations, vs
static fields to get a better idea if things are working correctly.

Still need to copy changes to the side-effecting U+F analysis.


Former-commit-id: 4df2d0820d4f7ce8eab396aa3ccfb0834ff34210