Fixed problem with any instance target: crashed if the type given to it was a IJavaFormalType.  Fixed this by making sure that the type given to it is never an IJavaFormalType.  This mean that I had to update elaborateUseExpression() and the handling of anonymous class expressions in EffectsVisitor to check for IJavaFormalType before calling it.  If the tyep is a formal type is converted to the least upper CLASS type. (Never an interface).
This should be safe with respect to the named region because you would
only be able to use regions known to be part of the type, and those
would all come from the upper bound.

 It makes the effect somewhat less specific, but I cannot imagine it
makes a huge difference in practice.  Once you have an AnyIntsanceTarget
out there, things are pretty bad any how.




Former-commit-id: 21788a629ce562973c69bdd33c6b0150f808de01