Updated analyses to deal with ReferenceObject and ValueObject annotations.
Added AnnotationTester classes for ReferenceObject and ValueObject. 
Fixed EqualityAnalysis to use ValueObjectAnnotationTester.

Updated ITypeFormalEnv and AnnotationBoundsTypeFormalEnv to handle
reference object and value object.

Updated GenericTypeInstantiationChecker to check for ReferneceObject and
ValueObject, and to handle the new OR semantics when a type formal is
named in more than one attribute.

Removed the old AnnotationBoundVirtualDrop.  

Still need to update the ThreadSafe, Containable, and Immutable anlaysis
to handle the OR semantics correctly.


Former-commit-id: 818c9beb4ec2ef8da84f61bbcada395bcb15a646