Updated TypeDeclAnnotationTester and its subclasses so that result folders from checking AnnotationsBounds on ParameterizedTypes are incorporated into the set of trusted results.
Updated ThreadSafeAnnotationTester, ContainableAnnotationTester, and
ImmutableAnnotationTester so have a flag that indicates whehter it tests
the type or the implementatino.  This allows the tester to be used in
{ThreadSafe, Containable, Immutable}Processor to test the type of new
expressions.

Updated Containable assurance so that a final field may be assigned new
Containable instance.  This makes the analysis like that of the
ThreadSafe and Immutable.


Former-commit-id: c61a46abe10a771783e93719c3571edca67d3a36