Fixed AnnotationBoundsTypeFormalEnv to use to the "allowReferenceObject" attribute of @Containable type annotations.
Fixed ParameterizedTypeAnalysis to add a trivial result drop when the
generic type is never actually parameterized.

Fixed ParameterizedTypeAnalysis to check annotation bounds correctly:
Was assuring the type "O<X>" below:

@AnnotationBounds(threadSafe="X", containable="X")
public class Test<X> {
        private O<X> f = null;
}

@AnnotationBounds(threadSafe="T", referenceObject="T")
class O<T> {
        // empty
}

but it should not have.  X is @ThreadSafe, or it is @Containable.  Class
O requires that T be @ThreadSafe or @ReferenceObject.  X cannot be given
to T because X might be @Containable and not @ThreadSafe or
@ReferenceObject.  (Things didn't get updated properly when we switched
from AND to OR semantics, I think.)

Actual type parameter needs to be tested against all possible type
annotations, and the actual is only acceptable if its set of type
annotations is a non-empty subset of the type formal's bounds.

Also fixed so that GenericTypeInstantiationChecker doesn't try to check
the bounds on type formals that are unbounded.


Former-commit-id: 0374dfbc664759d96f29afcae366bee91a7f8959