Added new unused classes that separate the XAnnotationTester functionality into different testers based on three different ways the testers are used: (1) To test if the type of a declared field satisfies the type annotation (2) To test if the type of a new expression assigned to a final field satisfies the type annotation (3) To test if the type of satisifies an annotation bound
THis is first step.  I'm hoping to continue to refactor this into a set
of three methods that are parameterized by an enum of the type
annotation.

The tester for threadsafe needs to be careful because for (3) we only
want to accept THREADSAFE, but for (1) and (2) we want ThreadSafe or
Immutable.  This affects the way annotation bounds are looked up in the
type foraml environemnt too.

I want to merge AnnotationBoundsTypeFormalEnv with
TypeDeclAnnotationTestser too.  This puts everytyhing in one place.


Former-commit-id: 1ba5fe1c650dbb584946456ff2544e16c7edc0d7