The UniquePromiseDrop is the only (non-scoped) promise drop to override isCheckedByAnalysis() which was messing with drop-sea.
The implementation in UniquePromiseDrop is

  @Override
  public boolean isCheckedByAnalysis() {
    if (isUniqueReturn) {
      return super.isCheckedByAnalysis();
    } else {
      return true;
    }
  }

I removed this override and the results look better, but for the life of
me, I have no idea why this would have been added. Aaron, Edwin. what
about @Unique might have made us put this into the code. It is playing
with if drop-sea uses T and red-dot by saying a promise is checked by
analysis when it may or may not be.

The base method checks if the promise is in XML and "lies" that it was
checked, if not (from source) it sees if it is intended to be checked by
analysis (like @Region would not be) and then checks for analysis
results about it.

Former-commit-id: cdea4aaf3e743022b6a9ee0669a14f4cec72d917