Fixed checkAssignability() in NonNullTypeChecker: Need to check whether the promise drop is one of the virtual @Nullable drops that we created. If so, we don't put positive results on it, only negative ones.  So we have to to test whether the promise drop is non-null and non-virtual. And in the else branch, we have to only create the virtual drop if it doesn't already exist.
The reason for this is we were seeing goofy results, under virtual
@Nullable promises, where all the postive results that might occur that
appear syntactically before a negative result are not given results in
the chain of evidence, but those that occur after a negative result do
appear.  In actuality we only want the negative results in this case.


Also: Started to simplify testChain()

Former-commit-id: 599ea7cf8ea658c831b866366b1418313b8aff17