Working on the "<field>:<lock>" lock name implemenation.
Works for synchronized() statements.  Doesn't work for 

   @RequiresLock("service:Lock")
   void callService() {
       service.doSomething();
   }

yet.  Different method is used to convert the requires lock frame into
held locks.



Updated LockUtils.convertObjectExpressionToCallerContext() to create a
NeededLock from <field>:<lock> lock names.  The object expression in
this case is a VariableDeclator, becuase I cannot generate a new
ThisExpression for the substitution.  This creates problems for the
HeldInstanceLock.mustSatisfy() method.

Updated HeldIntanceLock.mustSatisfy() to check for the special case that
the needed lock's "object" portion is a VariableDeclarator.  In that
case, we delegate to a new abstract isFieldExprOfThis() method. 
Implemented this for AASTHeldIntanceLock and IRHeldInstanceLock.  This
method checks that the held lock's object expression is of the form
"this.f", where f is the lock named in the VariableDeclarator of the
needed lock.


Former-commit-id: 693bfb80d5ef56429d382c78bf86ae2526db9520