Had created the LockSpecProcessor abstract class and NeededLockProcessor subclass and tested them.  But then I accidently overwrote them, and now I have to do it again.
Fixed problem with NonNullTypeChecker. The handleMethod and
handleConstructor methods need to reset the receiverDecl on exit.  This
is so that nested anonymous classes do not poison the meothd that
contains them.  That is, when a method of the anonymous class is exited,
the receiverDecl reverts back to the receiverDecl fo the method that
contains the anonymous class.  Otherwise the ThisExpressionBinder was
returning wacky results.

Former-commit-id: b40357c78100f62459eedb7ecf6cabf412564ca0