Updated AbstractJavaAnalysisDriver to skip visiting inside of annotations because annotations do not have control flow graph structure.  Thus, the tree has no control flow results for annotation elements.
Cannot imagine why we would need flow graph results for annotations
because they are limited to constant expressions that are better
analyzed syntactically.

This came up because of autoboxing, which can introduce toString()
methods calls that the NonNullTypeChecker wanted to check the receiver
of.

import com.surelogic.*;
import static test.TestConstExprsInAnnos.*;

@RegionLocks({
	@RegionLock("L"+_+"is"+_+me+" protects foo"+(1+1))
})
public class TestConstExprsInAnnos {
	static final char _ = ' ';
	static final String me = "this";
	int foo2;
}

In particular, the use of TestConstExprsInAnnos._ in "L"+_ is boxed to a
Character object and toSTring() is called on that object.



Former-commit-id: 39234a19a88dd3eb405c3c3cdfb12a97e9d8d569