Been finding problems trying to run Nullable assurance over AOSP.
(1) The flow graph didn't handle super and qualified super (C.super)
correctly in JavaEvaluationTransfer.transferUse().  Was calling
transferUseVar() with them instead of forwarding to
transferUseReceiver() and transferUseQualifiedReceiver().

(2) The flow graph didn't handle VarArgsExpressions.  Had to update
JavaEvaluationTransfer.transferOperation() to check for them and call a
new transferVarArgs() method.  Then I had to fix transferCall(final
IRNode node, T value, final Iterable<IRNode> actuals, final boolean
hasOuter, final boolean isMethodCall) to not treat VarArgs as a special
case because they are now handled correctly elsewhere in the flow.

(3) More problems with C.super.  This time, we have C.super inside the
class C so that it is refers to the same object as "this":

public class Outer {
    public class PrivateAccess extends Super {
        public int getValue() {
            return PrivateAccess.super.getValue();
        }
    }
}

class Super {
    public int getValue() {
        return 0;
    }
}

In this case, the NonNullRawTypeAnalysis is doing the right thing,
treating PrivateAccess.super as "this", but the NonNullTypeChecker is
failing in testChain() because its trying to use PrivateAccess.super as
the a variable instead of the ReceiverDeclaration.  I think the solution
is to fix the ThisExpressionBinder to bind the X.super correctly.  See
JLS 15.12.4, page 482: "If the fourth production for MethodInvocation,
ClassName . super, is involved, then the target reference is the value
of ClassName.this."  So I have to update the ThisExpressionBinder to
handle QualifiedSuperExpressions the same as QualifiedThisExpressiosn.

Fixed AbstractThisExpressionBinder.bindThisExpression_private() to have
a case for QualifiedSuperExpression.  It is identical to the case for
QualifiedThisExpression.


Former-commit-id: 961426a42581df983d15a59734ccb31f2b094351