Yesterday and today.  FIxed problem with GenericTypeInstantiationChecker.  Had problem with this
public class TestDiamond {
	void main() {
		Set set = new LinkedHashSet<>(3);
		set.add(new Object());
	}
}

Two issues: The <> operator, and fact that the inferred type of
LinkedHashSet<> is LinkedHashSet (raw).  

Edwin added a bind IBinder.getInferredJavaType() so that we can get the
inferred type fo the parameterized type (LinkedHashSet<> in this case).
I then had to add a check to see if the inferred type is actually
parameterized: if not, we stop checking the parameters because there
aren't any.   These fixe are in visitParameterizedType().  Also had to
fix checkActualsAgainstBounds() because it was trying to get parse nodes
of the actual parameters from the parameterized type node, but these
don't exist when the <> opeator is used.  In this case, I just use the
node of the diamond.

Former-commit-id: 478732d1be5d6215e5d00cbb83a4c6991f6a6c94