Implemented TrackPartiallyInitalized.
Scrubber: Only on class declarations, not interface declarations

Updated NullablePreprocessor to do the checks for
TrackPartiallyInitialized because it already uses a
JavaSemanticsVisitor.  NonNull fields are linked to TPI annotatiosn on
the class.  TPI annotation on a class is linked to the TPI annotation on
the super class (or the fact that the class is Object).   

Updated NonNullRawTypeAnalysis.Transfer.transferConstructorCall().
Only pushes a partially initialized reference on the stack if the class
is TPI.


Former-commit-id: 11664b4d9494ad6ef0ac44ab54d17196d4aab610