Changed EffectsVisitor to use more Guava ImmutableSets.
Added new message for AnonClassEvidence.  This replaces the "held as"
messages that we used to use.  This is now taken care of inside the
effects analysis and the AnonClassEvidene on a Target records when the
Qualified receiver is remapped.  Added evidence processor ot the lock
analysis to report this supporting information.



Former-commit-id: 8e6cc76c6bf450240d21c3b2c3742132d46a3f8f